Initial Database operations see description for more details
It will automatically create a database provided it a server is running. The file still needs to be modified to work with the rest of beat. Namely the lines that have to be changed are line 13 to take the name of the current project being worked on, line  15 to take the name of the current point of interest data being saved, line 61 to select the point of interest being saved, and lines 67-80 for the data being entered for the point of interest. This version has  methods for adding, updating/creating, getting, and removal of data, but only adding has been implemented.