Vars displayed differently, open radare in debug
Variables are displayed along with functions and show up as children
to their corresponding function. As a result, changed the way
the variable file is formatted and how variables are read. Radare is
opened in debug mode to allow for breakpoints. The dynamic run button
now updates the POI's currently being displayed in addition to running
dynamic analysis.